[KAN-26] Basic multi-threading implementation. Moved from grequests to a manual implementation. Currently, makes every request its own thread which is quite intensive on user hardware. Was able to return 50 requests in about 17 seconds. That request was: PWM (Portland, Maryland) -> KENV (Wendover, Utah). Any airports further than these two will not work, because of the FAA API request limit.
[KAN-26] Added a new error handler message in get_notams_at
for status code 429. This code occurs when the
user has reached their request limit.

[KAN-26] Changed thread checking to increase throughput.
Before the program would attempt to get the result
from each thread in the order that they were
located in the array.
Now, the program will only attempt to retrieve
values from threads that are finished, thus it no
longer wastes time waiting for a thread while
others are already finished.

[KAN-26] Fixed bugs in test files.
These were encountered when merging to main.